scripts/pbsep/pyproject.toml
[project]
name = "pbsep"
version = "1.0.0"
description = "PBSEP-256: Photorealistic to Binary Symbol Extraction Pipeline"
requires-python = ">=3.11"
dependencies = [
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "scikit-image>=0.21.0",
    "click>=8.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
saliency = [
    "torch>=2.0.0",
    "timm>=0.9.0",
]
dev = [
    "pytest>=7.0.0",
]

[project.scripts]
pbsep = "pbsep.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pbsep"]
